+++ packages/react-ui/components/Textarea/Textarea.styles.ts
@@ -35,7 +35,6 @@ export const getStyles = memoizeGetStyles(({ css }: Emotion) => ({
       background: ${t.textareaBg};
       background-clip: ${t.textareaBackgroundClip};
       border: ${t.textareaBorderWidth} solid ${t.textareaBorderColor};
-      border-top-color: ${t.textareaBorderTopColor};
       box-shadow: ${t.textareaShadow};
       box-sizing: border-box;
       color: ${t.textareaColor};