+++ packages/react-ui/components/TokenInput/TokenInput.tsx
@@ -557,8 +557,10 @@ export class TokenInput<T = string> extends React.PureComponent<TokenInputProps<
         ) : null}
       </>
     );
+
     const container = maxHeight ? (
       <ScrollContainer
+        className={this.styles.inputPlaceholderScrollWrapper()}
         scrollRef={this.scrollContainerRef}
         maxHeight={maxHeight}
         showScrollBar="always"