+++ packages/react-ui/internal/MobilePopup/MobilePopup.styles.ts
@@ -53,8 +53,8 @@ export const getJsStyles = memoizeGetStyles(({ css }: Emotion) => ({
       right: 0;
       left: 0;
       z-index: 9999;
+      touch-action: none;
       background: #333333;
-      pointer-events: none;
       opacity: 50%;
     `;
   },